+++ CMakeLists.txt
@@ -141,9 +141,9 @@ set(CMAKE_Fortran_MODULE_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}/include)
 set(SRC_FTI
 	src/api.c src/checkpoint.c src/conf.c src/meta.c
 	src/postckpt.c src/postreco.c src/recover.c
-	src/tools.c src/topo.c src/ftiff.c src/hdf5.c
-	src/diff-checkpoint.c src/stage.c src/incremental-checkpoint.c
-	src/failure-injection.c src/api_cuda.c src/utility.c)
+	src/tools.c src/topo.c src/ftiff-io.c src/hdf5-io.c
+	src/dcp.c src/stage.c src/icp.c
+	src/fi-io.c src/api-cuda.c src/utility.c)
 
 if (ENABLE_GPU)
   include_directories(${CUDA_INCLUDE_DIRS})
@@ -269,8 +269,8 @@ set(FTI_TARGETS fti.static fti.shared)
 install(TARGETS fti.static fti.shared DESTINATION lib)
 install(FILES "include/fti.h"
 	DESTINATION "include")
-install(DIRECTORY "include/fti-int"
-	DESTINATION "include" )
+install(FILES "include/fti-types.h"
+	DESTINATION "include")
 
 if(ENABLE_FORTRAN)
     CHECK_Fortran_COMPILER_FLAG("-cpp" FC_HAS_CPP_FLAG)
