+++ src/conf.c
@@ -36,9 +36,9 @@
  *  @brief  Configuration loading functions for the FTI library.
  */
 
-#include <fti-int/interface.h>
+#include "interface.h"
 
-#include "api_cuda.h"
+#include "api-cuda.h"
 
 /*-------------------------------------------------------------------------*/
 /**
