+++ src/ftiff-io.h
@@ -1,7 +1,7 @@
 #ifndef FTIFF_H
 #define FTIFF_H
 
-#include <fti-int/types.h>
+#include <fti.h>
 
 #define FTI_DCP_MODE_OFFSET 2000
 #define FTI_DCP_MODE_MD5 2001
