+++ src/hdf5-io.c
@@ -37,9 +37,9 @@
  */
 
 #ifdef ENABLE_HDF5
-#include <fti-int/interface.h>
+#include "interface.h"
 
-#include "api_cuda.h"
+#include "api-cuda.h"
 
 /*-------------------------------------------------------------------------*/
 /**
