+++ src/icp.c
@@ -1,5 +1,5 @@
 
-#include <fti-int/interface.h>
+#include "interface.h"
 
 /*-------------------------------------------------------------------------*/
 /**
