+++ src/meta.c
@@ -36,7 +36,7 @@
  *  @brief  Metadata functions for the FTI library.
  */
 
-#include <fti-int/interface.h>
+#include "interface.h"
 
 /*-------------------------------------------------------------------------*/
 /**
