+++ src/postckpt.c
@@ -36,7 +36,7 @@
  *  @brief  Post-checkpointing functions for the FTI library.
  */
 
-#include <fti-int/interface.h>
+#include "interface.h"
 
 /*-------------------------------------------------------------------------*/
 /**
