+++ src/topo.c
@@ -36,7 +36,7 @@
  *  @brief  Topology functions for the FTI library.
  */
 
-#include <fti-int/interface.h>
+#include "interface.h"
 
 /*-------------------------------------------------------------------------*/
 /**
