+++ src/utility.c
@@ -1,7 +1,7 @@
 #include <string.h>
 
-#include <fti-int/interface.h>
-#include "api_cuda.h"
+#include "interface.h"
+#include "api-cuda.h"
 
 
 /*-------------------------------------------------------------------------*/
