+++ test/local/keepL4Ckpt/Makefile.in
@@ -6,7 +6,7 @@ FTI_BUILD ?= @CMAKE_BINARY_DIR@
 FTI_RELEASE ?= @CMAKE_INSTALL_PREFIX@
 FTI_INC_DIR := $(FTI_RELEASE)/include
 FTI_LIB_DIR := $(FTI_RELEASE)/lib
-FTI_SRC := $(FTI_HOME)/src/*.h $(FTI_HOME)/src/*.c $(FTI_HOME)/include/fti.h
+FTI_SRC := $(FTI_HOME)/src/*.h $(FTI_HOME)/src/*.c $(FTI_HOME)/include/*.h
 WORK_DIR := $(FTI_HOME)/test/local/diffckpt
 
 .PHONY: clean all run fti
